Create README.md
Week-1-Assignment-Promineo-Tech
Instructions: In Eclipse, or an IDE of your choice, write the code that accomplishes the objectives listed below. Ensure that the code compiles and runs as directed. Take screenshots of the code and of the running program (make sure to get screenshots of all required functionality) and paste them in this document where instructed below. Create a new repository on GitHub for this week’s assignments and push this document, with your Java project code, to the repository. Add the URL for this week’s repository to this document where instructed and submit this document to your instructor when complete.

Coding Steps:

Create a new Java project in Eclipse (make sure the JRE is set to 1.8).
Create a new class in the project and name it App. Make sure the box is checked for the option that reads “public static void main(String[] args).
Inside the main method, create two variables for the following real-life examples and assign them values (choose the best data type for the values): a. Item price b. Amount of money in wallet c. Number of friends d. Age in years (as a whole number) e. First name f. Last name g. Middle initial
Create the following variables by performing operations (addition, subtraction, concatenation) on the variables created in the previous step: a. New amount of money in wallet after buying the item b. Number of friends you’ve made each year based on your age variable and your number of friends variable c. Full name based on first name, middle initial, and last name
Use System.out.println() to print out the values of all the variables you’ve created. Provide some detail as to what the value being printed is. For example, if I had a variable called favorite state, I would do the following: String favoriteState = “AZ”; System.out.println(“My favorite state is: “ + favoriteState);